backend/requirements.txt
fastapi
fastapi[standard]
motor
pydantic
pymongo
python-dotenv
uvicorn
